# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CEP Search tool.",
        "title": "Go Expert - CEP Timeout",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Diogo Silverio"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/cep/{cep}": {
            "get": {
                "description": "Fetch CEP/Address from two different providers - ViaCEP and BrasilAPI.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cep"
                ],
                "summary": "Get CEP/Address from provider",
                "parameters": [
                    {
                        "type": "string",
                        "description": "cep",
                        "name": "cep",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CEP found",
                        "schema": {
                            "$ref": "#/definitions/dto.AddressDTO"
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AddressDTO": {
            "type": "object",
            "properties": {
                "cep": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "neighborhood": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                }
            }
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}